refactor: merge internal/quictransport into raft/transport
The shared QUIC transport (Shared, ALPNListener, Listener) lived in a
separate internal/quictransport package. Merge it directly into
raft/transport so there is a single transport package for all QUIC
concerns — both the raft message transport and the shared UDP socket
multiplexing.

Updated importers:
- storage/engine.go
- storage/engine_test.go
- storage/cluster/cluster.go
- storage/cluster/transport.go
- raft/nodehost.go (de-duplicated import)

Removed: internal/quictransport/
